pyproject.toml
[project]
name = "actilabelv2"
version = "0.1.0"
description = "Wearable Annotation Tool for sensor data and images"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "actipy>=3.5.0",
    "matplotlib>=3.10.1",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "pillow>=11.2.1",
    "pygame>=2.6.1",
    "pyyaml>=6.0.2",
]

[project.scripts]
label_study = "actilabelv2.label_study:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
